Removes mapping ctrl-c and ctrl-v stuffs
We cannot change ctrl-c and ctrl-v maps because these are default vim
commands. Ctrl-v stands for blockwise visual selection and ctrl-c stands
for exit insert mode to normal mode.

Although we are used to use ctrl-c ctrl-v to copy and paste, we have to
adapt so many things to think in vim mode that those are just two more.