Refactor colorscheme code organization
This change removes the colors.vim config file because it handle
something that is personal. Following our guideline, everything personal
must be in my_config.vim. So, this orientation was put on README and
installation process was changed to not use this file anymore.

Now when user enables color support the installation script only add
colorscheme.vim to the vimrc file.